+++ tomcat/src/main/java/nextstep/jwp/controller/Controller.java
@@ -1,18 +1,18 @@
 package nextstep.jwp.controller;
 
-import org.apache.http.ContentType;
-import org.apache.http.HttpHeaders;
-import org.apache.http.HttpRequest;
-import org.apache.http.HttpResponse;
-import org.apache.http.HttpStatus;
-import org.apache.http.ViewResolver;
-import org.apache.http.annotation.RequestMapping;
+import org.apache.coyote.http11.ContentType;
+import org.apache.coyote.http11.HttpHeaders;
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.ViewResolver;
+import org.apache.coyote.http11.annotation.RequestMapping;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-@org.apache.http.annotation.Controller
+@org.apache.coyote.http11.annotation.Controller
 public final class Controller {
 
     @RequestMapping(method = "GET", path = "/")
