+++ tomcat/src/main/java/org/apache/coyote/RequestRouter.java
@@ -1,13 +1,13 @@
 package org.apache.coyote;
 
-import org.apache.http.ContentType;
-import org.apache.http.HttpHeaders;
-import org.apache.http.HttpRequest;
-import org.apache.http.HttpResponse;
-import org.apache.http.HttpStatus;
-import org.apache.http.ViewResolver;
-import org.apache.http.annotation.Controller;
-import org.apache.http.annotation.RequestMapping;
+import org.apache.coyote.http11.ContentType;
+import org.apache.coyote.http11.HttpHeaders;
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.ViewResolver;
+import org.apache.coyote.http11.annotation.Controller;
+import org.apache.coyote.http11.annotation.RequestMapping;
 
 import java.io.File;
 import java.io.IOException;
