+++ tomcat/src/main/java/org/apache/coyote/http11/ContentType.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.coyote.http11;
 
 public enum ContentType {
