+++ tomcat/src/main/java/org/apache/coyote/http11/HttpHeaders.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.coyote.http11;
 
 import java.util.HashMap;
 import java.util.Map;
