+++ tomcat/src/main/java/org/apache/coyote/http11/ViewResolver.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.coyote.http11;
 
 import java.io.File;
 import java.net.URL;
